+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.0
+
+- no longer requires you to open SDK panel
+- added [VRC Questifyer](https://github.com/imagitama/vrc-questifyer) 1.7+ integration
+
 # 1.2.0
 
 - filter avatars with [Quest] in their name